홈페이지에 축제 미리보기 추가

+++ app/src/main/java/com/amazing/stamp/pages/home/HomeFestivalsActivity.kt
@@ -81,7 +81,7 @@ class HomeFestivalsActivity : ParentActivity() {
         Log.d(TAG, "getFestival: ${KorTripInfoAPI.tripDateFormat.format(System.currentTimeMillis())}")
         val festivalCall = korTripInfoAPI.getFestivalInfoCall(
             SecretConstants.KOR_TRIP_INFO_SERVICE_KEY,
-            KorTripInfoAPI.tripDateFormat.format(System.currentTimeMillis()).toInt(),
+            KorTripInfoAPI.tripDateFormat.format(System.currentTimeMillis()),
             "json",
             "AND",
             "STAMF",
